+++ google_play_scraper/constants/google_play.py
@@ -3,5 +3,4 @@ from enum import Enum
 
 class Sort(int, Enum):
     NEWEST = 2
-    RATING = 3
     MOST_RELEVANT = 1